webui: move normalizeComparisonValue to utils/precision
Address review feedback from allozaur
